[tests] Factor out AuthorityState utils from authority_tests
- Move various utility functions into authority_test_utils to be used by the transactional test framework, which cannot exist in cfg(test)
